requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.3.1
enum34==1.1.6
Flask==1.0.2
Flask-MySQL==1.4.0
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
PyMySQL==0.9.2
six==1.11.0
Werkzeug==0.14.1
mysql_connector